requirements.txt
gym==0.26.2
matplotlib==3.10.5
numpy==2.3.2
pygame==2.6.1
pyserial==3.5
torch==2.5.1+cu118